Create ItemSpawnerManager.lua
if you can, Please check this against a vanilla ISM, I "Hard Modded" this ISM (directly edited the ISM instead of doing it the proper way).  The significant changes are the creation of the "RandomRangedWithMagsPure" category which is basically a copy/paste of the RandomRangedPure category but I made it so each weapon would spawn with atleast 2 magazines.  Next biggest change was all xxweaponWithMagazines/Rounds group now spawns 2 magazines instead of 1 and weapons that have extended magazines have a chance of spawning with its extended magazine.